py3 compatibility which requires new pefile
Due to the fact that the old pefile isn't compatible with Python 3 we
are required to upgrade it to the latest version. This also implies
backwards incompatibility with Cuckoo (as it has pefile pinned to
pefile2==1.2.11) and therefore requires roach to release a version
0.2 which would then be included in a new Cuckoo release that bumps
pefile (or, more precisely, pefile2).

Note: we probably need to mirror pefile into pefile2 as otherwise we'd
end up with environments that have both pefile and pefile2 installed,
potentially bricking it in a way we can't resolve through Cuckoo.